game state tracks if game is over

+++ main.py
@@ -18,6 +18,7 @@ def main():
     my_font = pygame.font.SysFont("monospace", 40)
     score = 0
     lives = 3
+    game_over = False
 
     print(f"Starting asteroids!\nScreen width: {SCREEN_WIDTH}\nScreen height: {SCREEN_HEIGHT}")
 
@@ -42,26 +43,29 @@ def main():
         for obj in updatable:
             obj.update(dt)
 
-        for asteroid in asteroids:
-            for shot in shots:
-                if asteroid.check_collision(shot):
-                    score += 10
-                    print(f"Score: {score}")
-                    asteroid.split()
-                    shot.kill()
 
-            if asteroid.check_collision(player):
-                player.kill()
-                lives -= 1
-                player = Player(SCREEN_WIDTH / 2, SCREEN_HEIGHT / 2)
-                updatable.add(player)
-                drawable.add(player)
-                print(f"Lives: {lives}")
-                print("You Died")
+        if not game_over:
 
-                if lives == 0:
-                    print("Game Over!")
-                    sys.exit()
+            for asteroid in asteroids:
+                for shot in shots:
+                    if asteroid.check_collision(shot):
+                        score += 10
+                        print(f"Score: {score}")
+                        asteroid.split()
+                        shot.kill()
+
+                if asteroid.check_collision(player):
+                    player.kill()
+                    lives -= 1
+                    player = Player(SCREEN_WIDTH / 2, SCREEN_HEIGHT / 2)
+                    updatable.add(player)
+                    drawable.add(player)
+                    print(f"Lives: {lives}")
+                    print("You Died")
+
+                    if lives == 0:
+                        game_over = True
+                        player.kill()
 
         score_text = f"Score: {score}"
         score_text_surface = my_font.render(score_text, True, (255, 255, 255))
@@ -78,11 +82,23 @@ def main():
         screen.blit(score_text_surface, (20,20))
         screen.blit(lives_text_surface, (20,SCREEN_HEIGHT - lives_text_surface.get_height() - 20))
 
+        if game_over:
+            player.kill()
+            gameover_text = "Game Over"
+            game_over_font = pygame.font.SysFont("monospace", 120)
+            gameover_text_surface = game_over_font.render(gameover_text, True, (255, 255, 255))
+            gameover_position = (
+                (SCREEN_WIDTH - gameover_text_surface.get_width()) / 2,
+                (SCREEN_HEIGHT - gameover_text_surface.get_height()) / 2
+            )
+            screen.blit(gameover_text_surface, gameover_position)
+
         pygame.display.flip()
 
 
         # framerate = 60 FPS
         dt = clock.tick(60) / 1000.0
+
